+++ R/exercise.R
@@ -1,5 +1,7 @@
 current_exercise_version <- "3"
 
+# Shiny Exercise Handling -------------------------------------------------
+
 # run an exercise and return HTML UI
 setup_exercise_handler <- function(exercise_rx, session) {
 
@@ -189,6 +191,8 @@ setup_exercise_handler <- function(exercise_rx, session) {
   })
 }
 
+# Validation and Standardization ------------------------------------------
+
 # This function exists to synchronize versions of the exercise objects in case
 # an exercise created with an older version of {learnr} is evaluated by a
 # newer version of {learnr}. This may be the case when there is a version
@@ -305,6 +309,8 @@ standardize_exercise_code <- function(exercise) {
   exercise
 }
 
+# Evaluate Exercise -------------------------------------------------------
+
 # evaluate an exercise and return a list containing output and dependencies
 # @param evaluate_global_setup - If `FALSE`, will not evaluate the global setup
 #   code. Instead, it just concatenates the exercise- specific setup code and
@@ -384,7 +390,7 @@ evaluate_exercise <- function(
   }
 
   # Check that user R code is parsable -------------------------------------
-  if (identical(tolower(exercise$engine), "r")) {
+  if (is_exercise_engine(exercise, "r")) {
     return_if_exercise_result(
       exercise_check_code_is_parsable(exercise)
     )
@@ -567,6 +573,8 @@ get_checker_func <- function(exercise, name, envir) {
   function(...) NULL
 }
 
+# Render Exercise ---------------------------------------------------------
+
 render_exercise <- function(exercise, envir) {
   # Protect global options and environment vars from modification by student
   local_restore_options_and_envvars()
@@ -586,10 +594,7 @@ render_exercise <- function(exercise, envir) {
   # Put the exercise in a minimal HTML doc
   output_format_exercise <- function(user = FALSE) {
     # start constructing knitr_options for the output format
-    knitr_options <- exercise$options
-    # Recreate the logic of `rmarkdown::knitr_options_html()` by setting these options
-    knitr_options$opts_chunk$dev <- "png"
-    knitr_options$opts_chunk$dpi <- 96
+    knitr_options <- exercise["opts_chunk"]
 
     if (isTRUE(user)) {
       knitr_options$knit_hooks$evaluate <- function(
@@ -633,10 +638,7 @@ render_exercise <- function(exercise, envir) {
 
   # Prepare code chunks containing exercise prep (setup) and user code
   rmd_src_prep <- exercise_code_chunks_prep(exercise)
-  rmd_src_user <- c(
-    readLines(system.file("internals", "templates", "exercise-setup.Rmd", package = "learnr")),
-    "", exercise_code_chunks_user(exercise)
-  )
+  rmd_src_user <- exercise_code_chunks_user_rmd(exercise)
 
   envir_prep <- duplicate_env(envir)
   # placeholder envir_result in case an error occurs with setup chunks
@@ -751,6 +753,24 @@ render_exercise <- function(exercise, envir) {
     )
   }
 
+  if (is_exercise_engine(exercise, "sql")) {
+    # make sql result available as the last value from the exercise
+    if (exists("___sql_result", envir = envir_result)) {
+      if (!is.null(exercise[["options"]][["output.var"]])) {
+        # the author expected the sql results in a specific variable
+        assign(exercise[["options"]][["output.var"]], last_value, envir = envir_result)
+      }
+      rm("___sql_result", envir = envir_result)
+    }
+
+    # make the connection object available in envir_prep (used by gradethis)
+    con_name <- exercise[["opts_chunk"]][["connection"]]
+    con <- get0(con_name, envir = envir, ifnotfound = NULL)
+    if (!is.null(con) && isS4(con) && inherits(con, "DBIConnection")) {
+      assign(con_name, con, envir = envir_prep)
+    }
+  }
+
   list(
     evaluate_result = evaluate_result,
     last_value = last_value,
@@ -760,26 +780,7 @@ render_exercise <- function(exercise, envir) {
   )
 }
 
-with_masked_env_vars <- function(code, env_vars = list(), opts = list()) {
-  # Always disable connect api keys and connect server info
-  env_vars$CONNECT_API_KEY <- ""
-  env_vars$CONNECT_SERVER <- ""
-  env_vars$LEARNR_EXERCISE_USER_CODE <- "TRUE"
-  # Hide shiny server sharedSecret
-  opts$shiny.sharedSecret <- ""
-
-  # Mask tutorial cache for user code evaluation
-  cache_current <- tutorial_cache_env$objects
-  tutorial_cache_env$objects <- NULL
-  withr::defer(tutorial_cache_env$objects <- cache_current)
-
-  # Disable shiny domain
-  shiny::withReactiveDomain(NULL, {
-    withr::with_envvar(env_vars, {
-      withr::with_options(opts, code)
-    })
-  })
-}
+# Exercise Chunk Helpers --------------------------------------------------
 
 exercise_get_chunks <- function(exercise, type = c("all", "prep", "user")) {
   type <- match.arg(type)
@@ -803,10 +804,7 @@ exercise_code_chunks_prep <- function(exercise) {
 }
 
 exercise_code_chunks_user <- function(exercise) {
-  # chunk options on the user chunk just duplicate the exercise$options
-  # which are set globally for the exercise
   user_chunk <- exercise_get_chunks(exercise, "user")
-  user_chunk[[1]]$opts <- NULL
   exercise_code_chunks(user_chunk)
 }
 
@@ -824,6 +822,28 @@ exercise_code_chunks <- function(chunks) {
   }, character(1))
 }
 
+
+exercise_code_chunks_user_rmd <- function(exercise) {
+  rmd_src_user <- c(
+    readLines(system.file("internals", "templates", "exercise-setup.Rmd", package = "learnr")),
+    "",
+    exercise_code_chunks_user(exercise)
+  )
+
+  if (is_exercise_engine(exercise, "sql")) {
+    rmd_src_user <- c(
+      rmd_src_user,
+      "",
+      '```{r eval=exists("___sql_result")}',
+      'get("___sql_result")',
+      "```"
+    )
+  }
+
+  rmd_src_user
+}
+
+
 exercise_get_blanks_pattern <- function(exercise) {
   exercise_blanks_opt <-
     exercise$options$exercise.blanks %||%
@@ -837,6 +857,8 @@ exercise_get_blanks_pattern <- function(exercise) {
 
   exercise_blanks_opt
 }
+
+# Exercise Check Helpers --------------------------------------------------
 
 exercise_check_code_for_blanks <- function(exercise) {
   blank_regex <- exercise_get_blanks_pattern(exercise)
@@ -1030,6 +1052,8 @@ exercise_highlight_unparsable_unicode <- function(code, pattern, line) {
   html_code_block(paste0(line, ": ", highlighted_code), escape = FALSE)
 }
 
+# Exercise Result Helpers -------------------------------------------------
+
 exercise_result_timeout <- function() {
   exercise_result_error(
     "Error: Your code ran longer than the permitted timelimit for this exercise.",
@@ -1112,6 +1136,12 @@ is_error_result <- function(x) {
   is_exercise_result(x) && length(x$error_message)
 }
 
+# Exercise Prep -----------------------------------------------------------
+
+is_exercise_engine <- function(exercise, engine) {
+  identical(knitr_engine(exercise$engine), knitr_engine(engine))
+}
+
 exercise_result_as_html <- function(x) {
   if (!is_exercise_result(x)) {
     return(NULL)
@@ -1150,50 +1180,107 @@ filter_dependencies <- function(dependencies) {
 
 
 prepare_exercise <- function(exercise) {
-  exercise$chunks <- lapply(exercise$chunks, function(chunk) {
-    isExercise <- identical(chunk$label, exercise$label)
-    chunk$opts <- merge_options(
-      preserved_opts = chunk$opts,
-      # don't include the exercise options in setup chunks
-      inherited_opts = if (isExercise) exercise$options else list(),
-      static_opts = if (isExercise) {
-        list(
-          eval = TRUE, echo = FALSE, tutorial = NULL,
-          cache = FALSE, child = NULL
-        )
-      } else {
-        # don't include results in setup chunks
-        list(include = FALSE)
-      }
-    )
-    # Move over user submission code to the pre-rendered chunk object
-    if (isExercise) {
-      chunk$code <- exercise$code
+  forced_opts_exercise <- list(
+    tutorial = NULL,
+    engine = NULL,
+    eval = TRUE,
+    echo = FALSE,
+    cache = FALSE,
+    child = NULL,
+    dev = "png",
+    dpi = 92
+  )
+
+  exercise <- prepare_exercise_if_sql(exercise)
+
+  exercise[["opts_chunk"]] <- merge_chunk_options(
+    inherited = exercise[["options"]],
+    forced = forced_opts_exercise
+  )
+
+  exercise$chunks <- lapply(exercise[["chunks"]], function(chunk) {
+    if (identical(chunk[["label"]], exercise[["label"]])) {
+      # Exercise Chunk ----
+      chunk[["opts"]] <- merge_chunk_options(
+        chunk = chunk[["opts"]],
+        inherited = I(exercise[["opts_chunk"]])
+      )
+      # keep only unique options that we over-rode when prepping specific ex type (e.g. sql)
+      different_ex_opt <- function(opt, name) !identical(opt, exercise[["opts_chunk"]][[name]])
+      chunk[["opts"]] <- chunk[["opts"]][imap_lgl(chunk[["opts"]], different_ex_opt)]
+      # move user submission code into the exercise chunk
+      chunk[["code"]] <- exercise[["code"]]
+    } else {
+      # Setup Chunk ----
+      chunk[["opts"]] <- merge_chunk_options(
+        chunk = chunk[["opts"]],
+        forced = list(include = FALSE, tutorial = NULL)
+      )
     }
     chunk
   })
+
+  # Restore opts_chunk to R list since merge_chunk_options() dputs the values
+  # These chunk options will be used as global knitr chunk options
+  exercise[["opts_chunk"]] <- lapply(
+    exercise[["opts_chunk"]],
+    function(opt) eval(parse(text = opt))
+  )
+  exercise[["opts_chunk"]] <- compact(exercise[["opts_chunk"]])
+
   exercise
 }
 
-# `preserved_opts` are options that user supplied in Rmd
-# `inherited_opts` are exercise options
-# `static_opts` are list of manually set options, e.g. list(include=FALSE) for setup chunks.
-merge_options <- function(preserved_opts, inherited_opts, static_opts = list()) {
+prepare_exercise_if_sql <- function(exercise) {
+  if (!is_exercise_engine(exercise, "sql")) {
+    return(exercise)
+  }
+
+  # Disable invisible warning (that's how sql chunks work)
+  exercise[["options"]][["exercise.warn_invisible"]] <- FALSE
+
+  # Set `output.var` so we can find it later, overwriting user name.
+  # After rendering, we'll remove the object or reassign it to the `output.var`
+  # the author was expecting. (This is much easier than dealing with knitr
+  # chunk options in various stages of evaluations/escaping.)
+  exercise[["chunks"]] <- lapply(exercise[["chunks"]], function(chunk) {
+    if (!identical(chunk[["label"]], exercise[["label"]])) {
+      return(chunk)
+    }
+    chunk[["opts"]][["output.var"]] <- "\"___sql_result\""
+    chunk
+  })
+
+  exercise
+}
+
+# `chunk` are options that user supplied in Rmd (assumed to be strings)
+# `inherited` are exercise options
+# `forced` are list of manually set options, e.g. list(include=FALSE) for setup chunks.
+merge_chunk_options <- function(
+  chunk = list(),
+  inherited = list(),
+  forced = list()
+) {
   # note: we quote each option's value if its type is a character, else return as is
   # to prevent rmd render problems (for e.g. fig.keep="high" instead of fig.keep=high)
-  static_opts <- lapply(static_opts, dput_to_string)
-  inherited_opts <- lapply(inherited_opts, dput_to_string)
+  if (!is_AsIs(forced)) {
+    forced <- lapply(forced, dput_to_string)
+  }
+  if (!is_AsIs(inherited)) {
+    inherited <- lapply(inherited, dput_to_string)
+  }
   # get all the unique names of the options
-  option_names <- unique(c(names(preserved_opts), names(inherited_opts), names(static_opts)))
+  option_names <- unique(c(names(chunk), names(inherited), names(forced)))
   opts <- lapply(option_names, function(option_name) {
     # first we want manually set options, then user's, then exercise
-    static_opts[[option_name]]  %||%
-      preserved_opts[[option_name]] %||%
-      inherited_opts[[option_name]]
+    forced[[option_name]]  %||%
+      chunk[[option_name]] %||%
+      inherited[[option_name]]
   })
   # since we manually grab the names, set the names to opts
   names(opts) <- option_names
-  # filter out options we don't need for the exercise.Rmd
+  # filter out options that already appear in the chunk item
   opts <- opts[!(names(opts) %in% c("label", "engine", "code"))]
   opts[!grepl("^exercise", names(opts))]
 }
@@ -1201,6 +1288,29 @@ merge_options <- function(preserved_opts, inherited_opts, static_opts = list()) 
 local_restore_options_and_envvars <- function(.local_envir = parent.frame()) {
   local_restore_options(.local_envir)
   local_restore_envvars(.local_envir)
+}
+
+# Exercise Eval Environment Helpers ---------------------------------------
+
+with_masked_env_vars <- function(code, env_vars = list(), opts = list()) {
+  # Always disable connect api keys and connect server info
+  env_vars$CONNECT_API_KEY <- ""
+  env_vars$CONNECT_SERVER <- ""
+  env_vars$LEARNR_EXERCISE_USER_CODE <- "TRUE"
+  # Hide shiny server sharedSecret
+  opts$shiny.sharedSecret <- ""
+
+  # Mask tutorial cache for user code evaluation
+  cache_current <- tutorial_cache_env$objects
+  tutorial_cache_env$objects <- NULL
+  withr::defer(tutorial_cache_env$objects <- cache_current)
+
+  # Disable shiny domain
+  shiny::withReactiveDomain(NULL, {
+    withr::with_envvar(env_vars, {
+      withr::with_options(opts, code)
+    })
+  })
 }
 
 local_restore_options <- function(.local_envir = parent.frame()) {
@@ -1227,127 +1337,7 @@ restore_envvars <- function(old) {
   do.call(Sys.setenv, as.list(old))
 }
 
-#' An Exercise Checker for Debugging
-#'
-#' An exercise checker for debugging that renders all of the expected arguments
-#' of the `exercise.checker` option into HTML. Additionally, this function is
-#' used in testing  of `evaluate_exercise()`.
-#'
-#' @param label Exercise label
-#' @param user_code Submitted user code
-#' @param solution_code The code in the `*-solution` chunk
-#' @param check_code The checking code that originates from the `*-check` chunk,
-#'   the `*-code-check` chunk, or the `*-error-check` chunk.
-#' @param envir_prep,envir_result The environment before running user code
-#'   (`envir_prep`) and the environment just after running the user's code
-#'   (`envir_result`).
-#' @param evaluate_result The return value from `evaluate::evaluate()`, called
-#'   on `user_code`
-#' @param last_value The last value after evaluating `user_code`
-#' @param engine The engine of the exercise chunk
-#' @param ... Not used (future compatibility)
-#'
-#' @keywords internal
-debug_exercise_checker <- function(
-  label,
-  user_code,
-  solution_code,
-  check_code,
-  envir_result,
-  evaluate_result,
-  envir_prep,
-  last_value,
-  engine,
-  ...
-) {
-  # Use I() around check_code to indicate that we want to evaluate the check code
-  checker_result <- if (is_AsIs(check_code)) {
-    local(eval(parse(text = check_code)))
-  }
-
-  tags <- htmltools::tags
-  collapse <- function(...) paste(..., collapse = "\n")
-
-  str_chr <- function(x) {
-    utils::capture.output(utils::str(x))
-  }
-
-  str_env <- function(env) {
-    if (is.null(env)) {
-      return("NO ENVIRONMENT")
-    }
-    vars <- ls(env)
-    names(vars) <- vars
-    x <- str_chr(lapply(vars, function(v) get(v, env)))
-    x[-1]
-  }
-
-  code_block <- function(value, engine = "r") {
-    tags$pre(
-      class = engine,
-      tags$code(collapse(value), .noWS = "inside"),
-      .noWS = "inside"
-    )
-  }
-
-  message <- htmltools::tagList(
-    tags$p(
-      tags$strong("Exercise label:"),
-      tags$code(label),
-      tags$br(),
-      tags$strong("Engine:"),
-      tags$code(engine)
-    ),
-    tags$p(
-      "last_value",
-      code_block(last_value)
-    ),
-    tags$details(
-      tags$summary("envir_prep"),
-      code_block(str_env(envir_prep))
-    ),
-    tags$details(
-      tags$summary("envir_result"),
-      code_block(str_env(envir_result))
-    ),
-    tags$details(
-      tags$summary("user_code"),
-      code_block(user_code, engine)
-    ),
-    tags$details(
-      tags$summary("solution_code"),
-      code_block(solution_code)
-    ),
-    tags$details(
-      tags$summary("check_code"),
-      code_block(check_code)
-    ),
-    tags$details(
-      tags$summary("evaluate_result"),
-      code_block(str_chr(evaluate_result))
-    )
-  )
-
-  list(
-    message = message,
-    correct = logical(),
-    type = "custom",
-    location = "replace",
-    checker_result = checker_result,
-    checker_args = list(
-      label           = label,
-      user_code       = user_code,
-      solution_code   = solution_code,
-      check_code      = check_code,
-      envir_result    = envir_result,
-      evaluate_result = evaluate_result,
-      envir_prep      = envir_prep,
-      last_value      = last_value,
-      engine          = engine,
-      "..."           = list(...)
-    )
-  )
-}
+# Print Methods -----------------------------------------------------------
 
 #' @export
 format.tutorial_exercise <- function (x, ...) {